+++ src/Game/Werewolf/Command/DevotedServant.hs
@@ -61,11 +61,11 @@ validateCommand callerName = do
 resetRole :: (MonadState Game m, MonadWriter [Message] m) => Text -> Role -> m ()
 resetRole callerName role
     | role == jesterRole            = jesterRevealed .= False
+    | role == orphanRole            = roleModel .= Nothing
     | role == simpleWerewolfRole    = do
         aliveWerewolfNames <- toListOf (players . werewolves . alive . name) <$> get
 
         tell $ devotedServantJoinedPackMessages callerName (aliveWerewolfNames \\ [callerName])
-    | role == wildChildRole         = roleModel .= Nothing
     | role == witchRole             = healUsed .= False >> poisonUsed .= False
     | role == wolfHoundRole         = allegianceChosen .= Nothing
     | otherwise                     = return ()